refactor(shm): extract SlotReader + move snapshot helpers into core/
Completes the SHM core/business split started by the SlotWriter
extraction. UnifiedReader now mirrors UnifiedWriter's composition
pattern:

  pub struct UnifiedReader {
      pub(crate) inner: SlotReader,
      channel_layouts: Vec<ChannelLayout>,
      channel_ids: Vec<u32>,
  }

`core::reader::SlotReader` owns the read-only mmap + slot_count +
max_slots and exposes header(), slot_at(), slot_count(), max_slots(),
writer_heartbeat(), is_writer_alive(), save_snapshot(). It implements
SlotIo (write_slot returns false, take_dirty_slots returns empty —
readers don't track dirty state). No business knowledge.

UnifiedReader's slot-level methods (slot_count, max_slots,
writer_heartbeat, is_writer_alive, save_snapshot) delegate to inner.
Channel-aware iterators (get_channel, get_instance, iter_*) stay as
inherent methods reading self.channel_layouts directly. The shared
`impl_shm_accessors!` macro is gone — its slot-level half now lives
on SlotReader/SlotWriter, and the channel-level half is two short
inherent methods on each adapter.

Also move snapshot serialization helpers from unified_shm.rs into
`core::snapshot_save`:

  - save_snapshot_impl (tear-resistant per-slot writer)
  - slot_snapshot_bytes (consistent-slot 32B encoder)
  - slot_unwritten_bytes (NaN sentinel 32B encoder)

These were already pure-infra (only need mmap bytes + slot_count +
PointSlot layout + SLOT_UNWRITTEN_BITS); placing them in core
finishes the snapshot path's decoupling. SlotWriter::save_snapshot
and SlotReader::save_snapshot both call into core::snapshot_save
directly.

After this commit, `core/` contains the entire pure-infra surface of
the SHM subsystem (slot, bitmap, config, header, slot_io, reader,
writer, snapshot_save). The crate can be promoted to a standalone
`voltage-shm-slots` mechanically — no further code surgery needed
to break the business coupling.

unified_shm.rs now houses only:
  - UnifiedWriter (channel-aware adapter over SlotWriter)
  - UnifiedReader (channel-aware adapter over SlotReader)
  - validate_shm_header / verify_file_min_size / verify_slot_count
    (open-time header validation — fine where they are)
  - reconfigure_existing / restore_from_snapshot (still mixed; their
    refactor is part of Step 3 atomic-swap design)

Verification
  - cargo check --workspace: pass
  - cargo clippy -p voltage-rtdb-shm -- -D warnings: clean
  - cargo fmt --check: clean
  - cargo test -p voltage-rtdb-shm: 85 tests pass
    (56 lib + 8 + 6 + 10 + 5 integration; 1 doc-test ignored)
