+++ libs/voltage-rtdb-shm/src/lib.rs
@@ -65,8 +65,10 @@ pub mod dispatch;
 
 // Re-exports for convenience
 pub use core::bitmap::{BitmapStats, SlotAllocation, SlotBitmap, SlotBitmapHeader};
+pub use core::reader::SlotReader;
 pub use core::slot::PointSlot;
 pub use core::slot_io::{SlotIo, SlotRead};
+pub use core::writer::SlotWriter;
 pub use instance_index::{DynamicInstanceLayout, InstanceIndex, SharedSlotRef};
 pub use notification::ShmNotification;
 #[cfg(unix)]
