+++ libs/voltage-rtdb-shm/src/unified_shm.rs
@@ -32,7 +32,7 @@ use crate::core::slot::PointSlot;
 use crate::layout::{ChannelLayout, allocate_layouts};
 use crate::shared_config::SharedConfig;
 use anyhow::{Context, Result, bail};
-use memmap2::{Mmap, MmapOptions};
+use memmap2::MmapOptions;
 use std::fs::{File, OpenOptions};
 use std::path::PathBuf;
 use std::sync::atomic::{AtomicU32, AtomicU64, Ordering, fence};
@@ -44,6 +44,7 @@ pub use crate::core::header::{
     slot_offset,
 };
 
+use crate::core::reader::SlotReader;
 use crate::core::writer::SlotWriter;
 
 fn read_ne_bytes<const N: usize>(buf: &[u8], start: usize, label: &str) -> Result<[u8; N]> {
@@ -66,134 +67,8 @@ fn read_u32_ne(buf: &[u8], start: usize, label: &str) -> Result<u32> {
     Ok(u32::from_ne_bytes(read_ne_bytes(buf, start, label)?))
 }
 
-// ========== Shared Helpers ==========
-
-/// Save shared memory data to a snapshot file (atomic write via temp + rename).
-///
-/// Shared implementation used by both `UnifiedWriter::save_snapshot` and
-/// `UnifiedReader::save_snapshot`.
-///
-/// # Tear-resistant slot serialization
-///
-/// Earlier versions did a raw `memcpy` of the mmap region. If the writer
-/// was mid-update through a seqlock at snapshot time, the snapshot
-/// captured torn bytes (new value + old raw, or seq=odd mid-write) and
-/// preserved them across restart — a stale or impossible reading would
-/// then be restored to SHM and propagate through Redis until overwritten
-/// by the next live write.
-///
-/// Now we iterate slots and call `try_load_consistent()` per slot. Torn
-/// reads (writer concurrently mid-update) produce `None` and we write
-/// the unwritten-NaN sentinel for that slot — restore treats it as
-/// never-written and skips it, matching the safe behavior of fresh
-/// slots. The header bytes are still copied verbatim (header atomics
-/// are byte-stable so a `memcpy` of a 64-byte aligned region is safe).
-pub(crate) fn save_snapshot_impl(
-    mmap_data: &[u8],
-    slot_count: usize,
-    path: &std::path::Path,
-    label: &str,
-) -> Result<()> {
-    use std::io::Write;
-
-    if mmap_data.len() < slot_offset() + slot_count * std::mem::size_of::<PointSlot>() {
-        bail!(
-            "snapshot source mmap too small: len={} slot_count={}",
-            mmap_data.len(),
-            slot_count
-        );
-    }
-
-    let temp_path = path.with_extension("tmp");
-    if let Some(parent) = path.parent() {
-        std::fs::create_dir_all(parent)
-            .with_context(|| format!("Failed to create snapshot directory: {:?}", parent))?;
-    }
-
-    let mut file = std::fs::File::create(&temp_path)
-        .with_context(|| format!("Failed to create temp snapshot file: {:?}", temp_path))?;
-
-    // Header: copied verbatim. AtomicU64 fields are byte-stable; this is
-    // a 64-byte region and individual atomic stores are torn-safe at the
-    // u64 level, so a single memcpy captures a consistent header.
-    file.write_all(&mmap_data[..slot_offset()])
-        .with_context(|| "Failed to write snapshot header")?;
-
-    // Slots: iterate one at a time via the seqlock-aware load. Torn slots
-    // become unwritten-NaN sentinels in the snapshot.
-    // SAFETY: bounds checked above; PointSlot is repr(C, align(32)).
-    let slots_ptr =
-        unsafe { mmap_data.as_ptr().add(slot_offset()) as *const crate::core::slot::PointSlot };
-    let mut torn = 0usize;
-    for i in 0..slot_count {
-        // SAFETY: i < slot_count, and the mmap covers slot_count slots.
-        let slot = unsafe { &*slots_ptr.add(i) };
-        let bytes = match slot.try_load_consistent() {
-            Some((value, raw, ts)) => slot_snapshot_bytes(value, raw, ts),
-            None => {
-                torn += 1;
-                slot_unwritten_bytes()
-            },
-        };
-        file.write_all(&bytes)
-            .with_context(|| "Failed to write snapshot slot")?;
-    }
-
-    file.flush().context("Failed to flush snapshot file")?;
-    file.sync_all().context("Failed to sync snapshot file")?;
-
-    std::fs::rename(&temp_path, path)
-        .with_context(|| format!("Failed to rename temp to snapshot: {:?}", path))?;
-
-    let data_size = slot_offset() + slot_count * std::mem::size_of::<PointSlot>();
-    if torn > 0 {
-        tracing::warn!(
-            "{} snapshot saved with {} torn slot(s) elided as unwritten: {:?}, size={} bytes, slots={}",
-            label,
-            torn,
-            path,
-            data_size,
-            slot_count
-        );
-    } else {
-        tracing::info!(
-            "{} snapshot saved: {:?}, size={} bytes, slots={}",
-            label,
-            path,
-            data_size,
-            slot_count
-        );
-    }
-    Ok(())
-}
-
-/// Encode a known-consistent slot as 32 bytes matching PointSlot's
-/// `#[repr(C)]` layout: value_bits(u64) | timestamp(u64) | raw_bits(u64) |
-/// seq(u32) | dirty(u32). seq is written as 2 (even, signaling "value
-/// committed") and dirty as 0. Native-endian bytes match the in-memory
-/// representation that restore will load directly.
-fn slot_snapshot_bytes(value: f64, raw: f64, ts: u64) -> [u8; 32] {
-    let mut out = [0u8; 32];
-    out[0..8].copy_from_slice(&value.to_bits().to_ne_bytes());
-    out[8..16].copy_from_slice(&ts.to_ne_bytes());
-    out[16..24].copy_from_slice(&raw.to_bits().to_ne_bytes());
-    out[24..28].copy_from_slice(&2u32.to_ne_bytes());
-    out[28..32].copy_from_slice(&0u32.to_ne_bytes());
-    out
-}
-
-/// Encode an unwritten-sentinel slot for the snapshot file: NaN
-/// value_bits + NaN raw_bits + seq=0, dirty=0. Matches the layout
-/// produced by `PointSlot::new()` and is recognized by restore via
-/// the NaN sentinel filter.
-fn slot_unwritten_bytes() -> [u8; 32] {
-    let mut out = [0u8; 32];
-    out[0..8].copy_from_slice(&crate::core::slot::SLOT_UNWRITTEN_BITS.to_ne_bytes());
-    // timestamp = 0 (already)
-    out[16..24].copy_from_slice(&crate::core::slot::SLOT_UNWRITTEN_BITS.to_ne_bytes());
-    // seq = 0, dirty = 0 (already)
-    out
-}
+// Snapshot serialization is now in `core::snapshot_save`; SlotWriter /
+// SlotReader call into it directly.
 
 /// Validate a shared memory header: checks magic, version, and routing hash
 ///
@@ -275,63 +150,12 @@ fn validate_reconfigurable_header(header: &UnifiedHeader) -> Result<u32> {
     Ok(header.max_slots)
 }
 
-/// Shared accessor methods for both UnifiedWriter and UnifiedReader.
-///
-/// Both types store an mmap region with identical layout: Header + PointSlot[].
-/// This macro generates `header()`, `slot_at()`, and `lookup()` to avoid duplication.
-macro_rules! impl_shm_accessors {
-    ($mmap_field:ident) => {
-        /// Get header reference
-        #[inline]
-        fn header(&self) -> &UnifiedHeader {
-            // SAFETY: mmap region starts with a valid UnifiedHeader.
-            // UnifiedHeader is #[repr(C, align(64))], mmap base is page-aligned.
-            unsafe { &*(self.$mmap_field.as_ptr() as *const UnifiedHeader) }
-        }
-
-        /// Get PointSlot at index
-        #[inline]
-        fn slot_at(&self, index: usize) -> &PointSlot {
-            assert!(
-                index < self.slot_count,
-                "slot_at: index {} out of bounds (slot_count={})",
-                index,
-                self.slot_count
-            );
-            // SAFETY: index is bounds-checked above. PointSlot is #[repr(C, align(32))].
-            unsafe {
-                let ptr = self.$mmap_field.as_ptr().add(slot_offset()) as *const PointSlot;
-                &*ptr.add(index)
-            }
-        }
-
-        /// Lookup slot by channel key
-        #[inline]
-        pub fn lookup(&self, channel_id: u32, point_type: u8, point_id: u32) -> Option<usize> {
-            self.channel_layouts
-                .get(channel_id as usize)?
-                .slot(point_type, point_id)
-        }
-
-        /// Get current slot count
-        #[inline]
-        pub fn slot_count(&self) -> usize {
-            self.slot_count
-        }
-
-        /// Get max slots
-        #[inline]
-        pub fn max_slots(&self) -> u32 {
-            self.max_slots
-        }
-
-        /// Get channel layouts
-        #[inline]
-        pub fn channel_layouts(&self) -> &[ChannelLayout] {
-            &self.channel_layouts
-        }
-    };
-}
+// The previous `impl_shm_accessors!` macro is gone. Slot-level accessors
+// (header, slot_at, slot_count, max_slots) now live on `SlotWriter` /
+// `SlotReader` in `core::writer` / `core::reader`. Channel-aware methods
+// (lookup, channel_layouts) live as explicit inherent methods on
+// `UnifiedWriter` / `UnifiedReader` and read `self.channel_layouts`
+// directly.
 
 // ========== Memory Layout ==========
 
@@ -1085,12 +909,10 @@ impl UnifiedWriter {
 /// Multiple readers allowed (modsrv, monarch, etc.).
 /// Builds indexes from ChannelPointCounts using same allocation algorithm.
 pub struct UnifiedReader {
-    mmap: Mmap,
-    max_slots: u32,
-    slot_count: usize,
-    /// Channel layouts (Vec index by channel_id)
+    pub(crate) inner: SlotReader,
+    /// Channel layouts (Vec index by channel_id) — business adapter state.
     channel_layouts: Vec<ChannelLayout>,
-    /// For monarch API: list of valid channel IDs
+    /// For monarch API: list of valid channel IDs — business adapter state.
     channel_ids: Vec<u32>,
 }
 
@@ -1137,9 +959,7 @@ impl UnifiedReader {
         );
 
         Ok(Self {
-            mmap,
-            max_slots,
-            slot_count,
+            inner: SlotReader::from_mmap(mmap, max_slots, slot_count),
             channel_layouts,
             channel_ids,
         })
@@ -1186,15 +1006,37 @@ impl UnifiedReader {
         );
 
         Ok(Self {
-            mmap,
-            max_slots,
-            slot_count,
+            inner: SlotReader::from_mmap(mmap, max_slots, slot_count),
             channel_layouts: Vec::new(),
             channel_ids: Vec::new(),
         })
     }
 
-    impl_shm_accessors!(mmap);
+    // Pure-infra accessors delegate to inner SlotReader.
+    #[inline]
+    pub fn slot_count(&self) -> usize {
+        self.inner.slot_count()
+    }
+    #[inline]
+    pub fn max_slots(&self) -> u32 {
+        self.inner.max_slots()
+    }
+    #[inline]
+    fn slot_at(&self, index: usize) -> &PointSlot {
+        self.inner.slot_at(index)
+    }
+    /// Channel-aware adapter: look up the slot index for a (channel, type, point).
+    #[inline]
+    pub fn lookup(&self, channel_id: u32, point_type: u8, point_id: u32) -> Option<usize> {
+        self.channel_layouts
+            .get(channel_id as usize)?
+            .slot(point_type, point_id)
+    }
+    /// Read-only access to the channel layout table — business adapter state.
+    #[inline]
+    pub fn channel_layouts(&self) -> &[ChannelLayout] {
+        &self.channel_layouts
+    }
 
     // ========== Point Query API ==========
 
@@ -1332,34 +1174,21 @@ impl UnifiedReader {
 
     // ========== Stats ==========
 
-    /// Get writer heartbeat
+    /// Writer heartbeat — delegates to inner SlotReader.
     #[inline]
     pub fn writer_heartbeat(&self) -> u64 {
-        self.header().writer_heartbeat.load(Ordering::Relaxed)
+        self.inner.writer_heartbeat()
     }
 
-    /// Check if writer is alive based on heartbeat timestamp
+    /// Check if writer is alive based on heartbeat timestamp — delegates.
     #[inline]
     pub fn is_writer_alive(&self, timeout_ms: u64) -> bool {
-        let heartbeat = self.writer_heartbeat();
-        if heartbeat == 0 {
-            return false;
-        }
-        let now = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap_or(std::time::Duration::ZERO)
-            .as_millis() as u64;
-        now.saturating_sub(heartbeat) < timeout_ms
+        self.inner.is_writer_alive(timeout_ms)
     }
 
-    // ========== Snapshot API ==========
-
-    /// Save current shared memory state to a snapshot file (read-only snapshot)
-    ///
-    /// This is useful for readers like modsrv to save state before shutdown.
-    /// Uses atomic write: writes to temp file first, then renames to final path.
+    /// Save current SHM state to a snapshot file — delegates to inner.
     pub fn save_snapshot(&self, path: &std::path::Path) -> Result<()> {
-        save_snapshot_impl(&self.mmap, self.slot_count, path, "Reader")
+        self.inner.save_snapshot(path)
     }
 }
